exp fit on X to calc spin gap is trash

+++ plotting/deleteData.py
@@ -4,19 +4,33 @@ import shutil
 
 N = sys.argv[1]
 
-print("removing old data (" + N + ") ...")
-for filename in os.listdir("results/3DData/" + N + "/C/"):
-    if filename == "dummyFile.txt":
-        continue
-    try:
-        shutil.rmtree("results/3DData/" + N + "/C/" + filename, ignore_errors=False, onerror=False)
-    except:
-        os.remove("results/3DData/" + N + "/C/" + filename)
+which =  sys.argv[2]
 
-for filename in os.listdir("results/3DData/" + N + "/X/"):
-    if filename == "dummyFile.txt":
-        continue
-    try:
-        shutil.rmtree("results/3DData/" + N + "/X/" + filename, ignore_errors=False, onerror=False)
-    except:
-        os.remove("results/3DData/" + N + "/X/" + filename)
+if which == "3D":
+    print("removing old 3D data (" + N + ") ...")
+    for filename in os.listdir("results/3DData/" + N + "/C/"):
+        if filename == "dummyFile.txt" or filename == "data_placeholder.txt":
+            continue
+        try:
+            shutil.rmtree("results/3DData/" + N + "/C/" + filename, ignore_errors=False, onerror=False)
+        except:
+            os.remove("results/3DData/" + N + "/C/" + filename)
+
+    for filename in os.listdir("results/3DData/" + N + "/X/"):
+        if filename == "dummyFile.txt":
+            continue
+        try:
+            shutil.rmtree("results/3DData/" + N + "/X/" + filename, ignore_errors=False, onerror=False)
+        except:
+            os.remove("results/" + N + "/X/" + filename)
+if which == "SG":
+    print("removing old spin gap data (" + N + ") ...")
+    for filename in os.listdir("results/" + N + "/data/spin_gap_data/"):
+        if filename == "dummyFile.txt" or filename == "data_placeholder.txt":
+            continue
+        try:
+            shutil.rmtree("results/" + N + "/data/spin_gap_data/" + filename, ignore_errors=False, onerror=False)
+        except:
+            os.remove("results/" + N + "/data/spin_gap_data/" + filename)
+else:
+    print("no valid data to delete")
